+++ lib/qlevar_router.dart
@@ -1,2 +1,3 @@
+export 'src/qpages.dart';
 export 'src/qr.dart';
 export 'src/types.dart';