+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## [0.3.0]
+
+- Fix Router update parent when it it component.
+- Add Page Types `QRPlatformPage`, `QRMaterialPage`, `QRCupertinoPage`,  `QRCustomPage` and `QRSlidePage`.
+- Better Tree Logging and method so the user can log it.
+
 ## [0.2.5] 24-01-2021
 
 - Set route info before onInit to use it when it needed.